reorg flask commands
